pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "thenine"
version = "0.1.0"
description = "Automated brand identity framework - generate complete brand packages from a single input"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Milan Pecka", email = "milan@dataworks.cz" },
]
dependencies = [
    "typer[all]>=0.15.0",
    "pydantic>=2.12.0",
    "coloraide>=4.0",
    "wcag-contrast-ratio>=0.9",
    "anthropic>=0.49.0",
    "cloudflare>=4.3.0",
    "weasyprint>=67.0",
    "build123d>=0.10.0",
    "trimesh>=4.11.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "Jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
]

[project.scripts]
thenine = "thenine.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["thenine"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"
